+++ frontend/src/Pages/LycheReservationPage/LycheReservationPage.tsx
@@ -1,22 +1,6 @@
-import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
-import { useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { z } from 'zod';
-import {
-  Button,
-  Checkbox,
-  Dropdown,
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-  Input,
-} from '~/Components';
-import type { DropdownOption } from '~/Components/Dropdown/Dropdown';
 import { Link } from '~/Components/Link/Link';
 import { SultenPage } from '~/Components/SultenPage';
 import { checkReservationAvailability } from '~/apis/sulten/sultenApis';
@@ -29,22 +13,8 @@ import {
   FindAvailableTablesForm,
   type FindTableData,
 } from './Components/FindAvailableTablesForm/FindAvailableTablesForm';
+import { ReservationDetailsForm, type ReservationFormData } from './Components/ReserveTableForm/ReserveTableForm';
 import styles from './LycheReservationPage.module.scss';
-
-// Combined schema for all form data
-const reservationSchema = z.object({
-  occasion: z.string().min(1, 'Required'),
-  guest_count: z.number().min(1).max(8),
-  reservation_date: z.date(),
-  start_time: z.string().min(1, 'Required'),
-  name: z.string().min(1, 'Required'),
-  phonenumber: z.string().min(1, 'Required'),
-  email: z.string().email('Invalid email address'),
-  additional_info: z.string().optional(),
-  agree: z.boolean().refine((val) => val === true, 'You must agree to the terms'),
-});
-
-type ReservationFormData = z.infer<typeof reservationSchema>;
 
 export function LycheReservationPage() {
   const { t } = useTranslation();
@@ -54,29 +24,6 @@ export function LycheReservationPage() {
   const [availableTimes, setAvailableTimes] = useState<AvailableTimes[]>([]);
   const [error, setError] = useState<string | null>(null);
   const [findTableData, setFindTableData] = useState<FindTableData | null>(null);
-  const termsText = useTextItem(TextItem.sulten_reservation_policy);
-
-  // Form for reservation details
-  const reservationForm = useForm<ReservationFormData>({
-    resolver: zodResolver(reservationSchema),
-    defaultValues: {
-      start_time: '',
-      name: '',
-      phonenumber: '',
-      email: '',
-      additional_info: '',
-      agree: false,
-    },
-  });
-
-  // Update default values of form when findTableData changes with useEffect
-  useEffect(() => {
-    if (findTableData) {
-      reservationForm.reset({
-        ...findTableData,
-      });
-    }
-  }, [findTableData, reservationForm]);
 
   // Use TanStack Query mutation for API call
   const checkAvailabilityMutation = useMutation({
@@ -131,120 +78,6 @@ export function LycheReservationPage() {
     // Here you would submit the data to your backend
   }
 
-  // Generate time options from available times returned by the API
-  const timeOptions: DropdownOption<string>[] = availableTimes
-    .filter((timeSlot: AvailableTimes) => timeSlot) // Filter out any empty values
-    .map((timeSlot: AvailableTimes) => ({
-      value: timeSlot as unknown as string,
-      label: timeSlot as unknown as string,
-    }));
-
-  // Reservation details form
-  const reservationDetailsForm = (
-    <Form {...reservationForm}>
-      <form onSubmit={reservationForm.handleSubmit(onReservationSubmit)} className={styles.formContainer}>
-        <div className={styles.reservation_info}>
-          <p className={styles.text}>
-            {t(KEY.common_date)} {findTableData?.reservation_date?.toLocaleDateString()}
-          </p>
-          <p className={styles.text}>
-            {t(KEY.common_guests)} {findTableData?.guest_count}
-          </p>
-        </div>
-
-        <FormField
-          control={reservationForm.control}
-          name="start_time"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>{`${t(KEY.common_time)}*`}</FormLabel>
-              <FormControl>
-                <Dropdown options={timeOptions} {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={reservationForm.control}
-          name="name"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>{`${t(KEY.common_name)}*`}</FormLabel>
-              <FormControl>
-                <Input type="text" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={reservationForm.control}
-          name="phonenumber"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>{`${t(KEY.common_phonenumber)}*`}</FormLabel>
-              <FormControl>
-                <Input type="text" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={reservationForm.control}
-          name="email"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>{`${t(KEY.common_email)}*`}</FormLabel>
-              <FormControl>
-                <Input type="email" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={reservationForm.control}
-          name="additional_info"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>{t(KEY.common_message)}</FormLabel>
-              <FormControl>
-                <Input type="text" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={reservationForm.control}
-          name="agree"
-          render={({ field }) => (
-            <FormItem>
-              <div className={styles.check_box}>
-                <FormControl>
-                  <Checkbox checked={field.value} onChange={field.onChange} />
-                </FormControl>
-                <FormLabel>{termsText}*</FormLabel>
-              </div>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <Button type="submit" theme="green">
-          {t(KEY.sulten_reservation_form_find_times)}
-        </Button>
-      </form>
-    </Form>
-  );
-
   return (
     <SultenPage>
       <div className={styles.container}>
@@ -263,8 +96,17 @@ export function LycheReservationPage() {
 
         {checkAvailabilityMutation.isPending && <div className={styles.loading}>{t('Checking availability...')}</div>}
 
-        {!checkAvailabilityMutation.isPending &&
-          (availableDate ? reservationDetailsForm : <FindAvailableTablesForm onSubmit={onFindTableSubmit} />)}
+        {!checkAvailabilityMutation.isPending && availableDate && findTableData && (
+          <ReservationDetailsForm
+            findTableData={findTableData}
+            availableTimes={availableTimes}
+            onSubmit={onReservationSubmit}
+          />
+        )}
+
+        {!checkAvailabilityMutation.isPending && !availableDate && (
+          <FindAvailableTablesForm onSubmit={onFindTableSubmit} />
+        )}
       </div>
     </SultenPage>
   );
